backend/requirements.txt
django
djangorestframework
django-cors-headers
PyJWT
PyPDF2
transformers
torch
scikit-learn
python-dotenv
sentence-transformers
